# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AwsConnection": {
        "description": "What the last Quick setup link and its stack reported, for the UI to follow live.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "disconnected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disconnected At"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "pending": {
            "default": false,
            "title": "Pending",
            "type": "boolean"
          },
          "plan_role_arn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Role Arn"
          },
          "reported_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported At"
          },
          "requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "role_arn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Arn"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "stack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Id"
          },
          "status": {
            "enum": [
              "waiting",
              "connected",
              "expired",
              "disconnected"
            ],
            "title": "Status",
            "type": "string"
          },
          "verification": {
            "anyOf": [
              {
                "enum": [
                  "pending",
                  "verified",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification"
          },
          "verification_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Error"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "status"
        ],
        "title": "AwsConnection",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The `\"detailed\"` error envelope every handler renders, and what OpenAPI advertises.\n\n`details` is present only on a 422, where it holds one entry per offending field.",
        "properties": {
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field validation failures. Present only on a 422.",
            "title": "Details"
          },
          "error_code": {
            "description": "A stable machine readable code, such as `NOT_FOUND`.",
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "description": "A human readable summary safe to surface to a caller.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "description": "The request id, echoed in the `X-Request-ID` header.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "The HTTP status code, repeated in the body.",
            "title": "Status",
            "type": "integer"
          },
          "success": {
            "default": false,
            "description": "Always `false` on an error response.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "status",
          "message",
          "request_id",
          "error_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Project": {
        "description": "A project and how many workspaces it holds.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "pattern": "^prj-(default|[0-9A-HJKMNP-TV-Z]{26})$",
            "title": "Project Id",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "workspace_count": {
            "default": 0,
            "title": "Workspace Count",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "name"
        ],
        "title": "Project",
        "type": "object"
      },
      "ProjectCreate": {
        "additionalProperties": false,
        "description": "A new project. Its name is unique across the environment, ignoring case.",
        "properties": {
          "description": {
            "default": "",
            "maxLength": 256,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 40,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]*$",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectList": {
        "description": "Every project, the default first and the rest by name.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProjectList",
        "type": "object"
      },
      "ProjectUpdate": {
        "additionalProperties": false,
        "description": "A partial project edit. A null description clears it; the name cannot be cleared.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ProjectUpdate",
        "type": "object"
      },
...
      "RunRoleSetup": {
        "description": "What a person needs to build the run role for one workspace.\n\nThe role cannot exist before the workspace does: its trust policy names the\nworkspace id as the external id, so the id has to be handed out first. Every\nworkspace response carries these three values so the setup can be followed\nwithout reading the stack's outputs. The control plane's credential vending\nrole is the only principal the trust policy may name: it assumes the role on\neach phase's behalf, so no runner task holds a path to it.",
        "properties": {
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "principal_arn": {
            "title": "Principal Arn",
            "type": "string"
          },
          "principal_arns": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Principal Arns",
            "type": "array"
          },
          "role_name": {
            "title": "Role Name",
            "type": "string"
          }
        },
        "required": [
          "principal_arn",
          "external_id",
          "role_name"
        ],
        "title": "RunRoleSetup",
        "type": "object"
      },
...
      "ValidationErrorDetail": {
        "description": "One offending field in a 422, carrying its location and reason but never its input.",
        "properties": {
          "field": {
            "description": "Dotted path to the offending field, or `_root` for the body itself.",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Why the value was rejected.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "The pydantic error type, such as `missing` or `string_type`.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ValidationErrorDetail",
        "type": "object"
      },
...
      "Workspace": {
        "description": "A stored workspace, with everything the run role setup needs.",
        "properties": {
          "auto_apply": {
            "default": false,
            "title": "Auto Apply",
            "type": "boolean"
          },
          "aws_connection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AwsConnection"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "engine": {
            "default": "terraform",
            "enum": [
              "terraform",
              "tofu"
            ],
            "title": "Engine",
            "type": "string"
          },
          "engine_version": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Engine Version",
            "type": "string"
          },
          "file_triggers_enabled": {
            "default": true,
            "title": "File Triggers Enabled",
            "type": "boolean"
          },
          "global_remote_state": {
            "default": false,
            "title": "Global Remote State",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pending_run_role_arn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Run Role Arn"
          },
          "plan_assume_role_arns": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Plan Assume Role Arns",
            "type": "array"
          },
          "plan_role_arn": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Role Arn"
          },
          "plan_secret_arns": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Plan Secret Arns",
            "type": "array"
          },
          "project_id": {
            "default": "prj-default",
            "title": "Project Id",
            "type": "string"
          },
          "remote_state_consumer_ids": {
            "items": {
              "type": "string"
            },
            "title": "Remote State Consumer Ids",
            "type": "array"
          },
          "run_api_token_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Run Api Token Scopes",
            "type": "array"
          },
          "run_role_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Role Account Id"
          },
          "run_role_arn": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Role Arn"
          },
          "run_role_checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Role Checked At"
          },
          "run_role_reconnect_required": {
            "default": false,
            "title": "Run Role Reconnect Required",
            "type": "boolean"
          },
          "run_role_setup": {
            "$ref": "#/components/schemas/RunRoleSetup"
          },
          "speculative_plans": {
            "default": true,
            "title": "Speculative Plans",
            "type": "boolean"
          },
          "tracked_branch": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracked Branch"
          },
          "trigger_patterns": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Trigger Patterns",
            "type": "array"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "vcs_installation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vcs Installation Id"
          },
          "vcs_repo": {
            "anyOf": [
              {
                "maxLength": 140,
                "pattern": "^[A-Za-z0-9-]+/[A-Za-z0-9._-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vcs Repo"
          },
          "vcs_repository_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vcs Repository Id"
          },
          "working_directory": {
            "default": "",
            "title": "Working Directory",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "engine_version",
          "workspace_id",
          "name",
          "created_at",
          "run_role_setup"
        ],
        "title": "Workspace",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Workspaces, variables, config versions and runs, as the gateway routes them.",
    "title": "WebbPulse Terraform control plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "description": "Every project with its workspace count, the default project first.",
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "List Projects",
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "description": "Create a project. The name has to be free, ignoring case.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Create Project",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "description": "Delete an empty project. Move its workspaces out first; nothing is moved for you.",
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^prj-(default|[0-9A-HJKMNP-TV-Z]{26})$",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Delete Project",
        "tags": [
          "workspaces"
        ]
      },
      "get": {
        "description": "One project with its workspace count. `prj-default` is the default project.",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^prj-(default|[0-9A-HJKMNP-TV-Z]{26})$",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Get Project",
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "description": "Rename a project or change its description. The default project is refused.",
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^prj-(default|[0-9A-HJKMNP-TV-Z]{26})$",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Update Project",
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspace_id}": {
      "delete": {
        "description": "Delete one workspace with its finished runs, current state and variables.\n\nA safe delete is a 409 carrying `WORKSPACE_MANAGES_RESOURCES` while the current\nstate tracks any resource instance: queue a destroy plan and apply it first, or\npass `force=true` to delete anyway and leave those resources unmanaged. Either\nmode is a 409 carrying `WORKSPACE_HAS_ACTIVE_RUN` while a run on the workspace\nhas not finished. Nothing is removed on a refusal.",
        "operationId": "delete_workspace_api_v1_workspaces__workspace_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "description": "Skip the managed resources check and delete even though state still tracks resources.",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "description": "Skip the managed resources check and delete even though state still tracks resources.",
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Delete Workspace",
        "tags": [
          "workspaces"
        ]
      },
      "get": {
        "description": "One workspace by id, with its run role setup.",
        "operationId": "get_workspace_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "description": "Edit one workspace. The name and the id are not editable.\n\nThe body is JSON Merge Patch: an omitted key leaves the stored value exactly\nas it was, and an explicit null on any field but `engine` and `engine_version`\nclears it. `model_dump(exclude_unset=True)` is what keeps the two apart, so a\nfield is only touched when the request carried its key.\n\nChanging or clearing `run_role_arn` drops the recorded check outcome, so the\nrole reads as unchecked until `run-role/check` says otherwise.\n\nConnecting another `vcs_repo` resolves it through the GitHub App the way the\ncreate does, and a null disconnects the repository.\n\nChanging a run role, the plan role or the reader roles is the workspace's AWS\nconnection, so a person has to have signed in within the step-up window; resending\nthe stored value is not a change.\n\nChanging `plan_role_arn`, `plan_assume_role_arns` or `plan_secret_arns` decides what\nevery plan, a pull request plan included, may reach, so it takes `admin`, or a run\ntoken holding the factory grant, as well as the step-up. The reader roles and the\nplan secrets together must fit the plan's session policy, or the edit is a 422\n`PLAN_SESSION_POLICY_TOO_LARGE`.\n\nChanging `run_api_token_scopes` hands every later run of this workspace a key on\nthis API, so it takes `admin` as well as the step-up.\n\nChanging `auto_apply` turns `runs:write` into the power to apply, so it takes\n`admin` and is recorded under `workspaces.workspace.auto_apply`. Turning it on also\ntakes the step-up, since it hands every later run's confirmation to the system.\n\n`global_remote_state` and `remote_state_consumer_ids` decide which other workspaces'\nruns may read this workspace's non-sensitive outputs. Every named consumer has to\nexist, or the edit is a 422 `REMOTE_STATE_CONSUMER_NOT_FOUND`.",
        "operationId": "update_workspace_api_v1_workspaces__workspace_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Update Workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
...
  }
}